backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==0.63.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.1
django-cors-headers==4.4.0
django-filter==24.3
django-bulk-update-or-create==0.3.0